Add unit tests to the disk cache to verify that a failure during cache reinitialization is handled properly.
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@137 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/disk_cache/disk_cache_test_base.cc
@@ -52,7 +52,8 @@ void DiskCacheTestBase::InitCache() {
     InitDiskCache();
 
   ASSERT_TRUE(NULL != cache_);
-  ASSERT_EQ(0, cache_->GetEntryCount());
+  if (first_cleanup_)
+    ASSERT_EQ(0, cache_->GetEntryCount());
 }
 
 void DiskCacheTestBase::InitMemoryCache() {
@@ -73,7 +74,8 @@ void DiskCacheTestBase::InitMemoryCache() {
 
 void DiskCacheTestBase::InitDiskCache() {
   std::wstring path = GetCachePath();
-  ASSERT_TRUE(DeleteCache(path.c_str()));
+  if (first_cleanup_)
+    ASSERT_TRUE(DeleteCache(path.c_str()));
 
   if (!implementation_) {
     cache_ = disk_cache::CreateCacheBackend(path, force_creation_, size_);
@@ -123,4 +125,9 @@ void DiskCacheTestBase::SimulateCrash() {
   if (size_)
     cache_impl_->SetMaxSize(size_);
   ASSERT_TRUE(cache_impl_->Init());
+}
+
+void DiskCacheTestBase::SetTestMode() {
+  ASSERT_TRUE(implementation_ && !memory_only_);
+  cache_impl_->SetUnitTestMode();
 }
